Remove reference to :keybindings command
The :keybindings command is not supported in recent versions of the Scala repl. It would be nice to explain some other way to get that same information here, but I don't know if there is anything comparatively simple. For now I think this text should just go away, so that people aren't scratching their head wondering why it doesn't work.