implement Snippet mechanism like Paradox
This implements low-tech solution for compiled documentation, inspired by Lightbend Paradox's snippet feature.
It's low tech because this feature itself just grabs the referenced file as code block and includes it, and asusmes the compiles validation is done elsewhere. This approach is nice because it's fast.

```
// This includes the entire file as Scala code snippet
@@snip [build.sbt]($root$/src/sbt-test/ref/basic/build.sbt) {}

or

// This includes snippet between a line containing #example another line with #example
@@snip [build.sbt]($root$/src/sbt-test/ref/basic/build.sbt) { #example }

or

// This specifies syntax highlight
@@snip [build.sbt]($root$/src/sbt-test/ref/basic/build.sbt) { #example type=text }
```

There's a hard-coded treatment for `$root$`, which is interpretted to be the root of the build. Otherwise, the path is treated to the relative path from the markdown file.