Various suggested changes
The section on `set ThisBuild/scalaVersion` right after the
`++2.13.0-M3!` section can be a bit confusing for a newcomer.
The user is left wondering: how is this different? What is
ThisBuild? What does that do differently, since the previous
example already changed the Scala version? Maybe it is better
to omit the previous section, or to postpone it until later.

Other than that, I added a bunch of missing articles, fixed
typos, and added some short explicative texts here and there,
thinking about what a person with no prior knowledge of sbt
might find confusing.