Rewrite triggered execution
This updates the Triggered Execution documentation to use the current
apis as of sbt 1.3.0. I still need to add a separate doc that explains
globs and incremental task evaluation using fileInputs. I will expand
the introduction to link to that doc to explain that inputs are
automatically monitored via those mechanisms.