docs: add a link to overriding the publishing convention
I was helping out someone today and assumed I'd find reference to
`CrossVersion` in the publishing section since they were trying to
publish a java artifact and looking for `CrossVersion.disabled`. However
it's in the cross-build page. This adds a small note in the publishing
page and then links it to the more detailed version in the cross build
page.